// app/src/data/schema.ts
export type Attribute =
  | 'Vaccine' | 'Data' | 'Virus' | 'Free' | 'Variable' | 'Unknown' | 'No Data';